+++ statsmodels/sandbox/gam_gsoc2015/gam.py
@@ -1,82 +1,7 @@
-import patsy
-from patsy import dmatrices, dmatrix, demo_data
-import pandas as pd
 import numpy as np
-import matplotlib.pyplot as plt
-from statsmodels.tools.numdiff import approx_fprime
-from statsmodels.api import GLM
-from scipy.interpolate import splev
-import scipy as sp
-from numpy.linalg import norm
-from statsmodels.discrete.discrete_model import Poisson, Logit, Probit
-from statsmodels.base.model import LikelihoodModel
-from statsmodels.robust.robust_linear_model import RLM       
        
-### Obtain b splines from patsy ###
 
-def _R_compat_quantile(x, probs):
-    #return np.percentile(x, 100 * np.asarray(probs))
-    probs = np.asarray(probs)
-    quantiles = np.asarray([np.percentile(x, 100 * prob)
-                            for prob in probs.ravel(order="C")])
-    return quantiles.reshape(probs.shape, order="C")
-
-
-
-## from patsy splines.py
-def _eval_bspline_basis(x, knots, degree):
-    try:
-        from scipy.interpolate import splev
-    except ImportError: # pragma: no cover
-        raise ImportError("spline functionality requires scipy")
-    # 'knots' are assumed to be already pre-processed. E.g. usually you
-    # want to include duplicate copies of boundary knots; you should do
-    # that *before* calling this constructor.
-    knots = np.atleast_1d(np.asarray(knots, dtype=float))
-    assert knots.ndim == 1
-    knots.sort()
-    degree = int(degree)
-    x = np.atleast_1d(x)
-    if x.ndim == 2 and x.shape[1] == 1:
-        x = x[:, 0]
-    assert x.ndim == 1
-    # XX FIXME: when points fall outside of the boundaries, splev and R seem
-    # to handle them differently. I don't know why yet. So until we understand
-    # this and decide what to do with it, I'm going to play it safe and
-    # disallow such points.
-    if np.min(x) < np.min(knots) or np.max(x) > np.max(knots):
-        raise NotImplementedError("some data points fall outside the "
-                                  "outermost knots, and I'm not sure how "
-                                  "to handle them. (Patches accepted!)")
-    # Thanks to Charles Harris for explaining splev. It's not well
-    # documented, but basically it computes an arbitrary b-spline basis
-    # given knots and degree on some specificed points (or derivatives
-    # thereof, but we don't use that functionality), and then returns some
-    # linear combination of these basis functions. To get out the basis
-    # functions themselves, we use linear combinations like [1, 0, 0], [0,
-    # 1, 0], [0, 0, 1].
-    # NB: This probably makes it rather inefficient (though I haven't checked
-    # to be sure -- maybe the fortran code actually skips computing the basis
-    # function for coefficients that are zero).
-    # Note: the order of a spline is the same as its degree + 1.
-    # Note: there are (len(knots) - order) basis functions.
-    n_bases = len(knots) - (degree + 1)
-    basis = np.empty((x.shape[0], n_bases), dtype=float)
-    der1_basis = np.empty((x.shape[0], n_bases), dtype=float)    
-    der2_basis = np.empty((x.shape[0], n_bases), dtype=float)    
-    
-    for i in range(n_bases):
-        coefs = np.zeros((n_bases,))
-        coefs[i] = 1
-        basis[:, i] = splev(x, (knots, coefs, degree))
-        der1_basis[:, i] = splev(x, (knots, coefs, degree), der=1)
-        der2_basis[:, i] = splev(x, (knots, coefs, degree), der=2)
-        
-
-    return basis, der1_basis, der2_basis
-
-
-
+## this class will be later removed and taken from another push
 class PenalizedMixin(object):
     """Mixin class for Maximum Penalized Likelihood
     TODO: missing **kwds or explicit keywords
@@ -216,7 +141,7 @@ class PenalizedMixin(object):
             else:
                 return res
 
-
+## this class will be later removed and taken from another push
 class Penalty(object):
     """
     A class for representing a scalar-value penalty.
@@ -237,7 +162,7 @@ class Penalty(object):
         """
         A penalty function on a vector of parameters.
         Parameters
-        ----------
+x        ----------
         params : array-like
             A vector of parameters.
         Returns
@@ -296,65 +221,12 @@ class GamPenalty(Penalty):
         return 2 * self.alpha * self.cov_der2
 
 
-class L2(Penalty):
-    """
-    The L2 (ridge) penalty.
-    """
-
-    def __init__(self, wts=None):
-        if wts is None:
-            self.wts = 1.
-        else:
-            self.wts = wts
-        self.alpha = 1.
-
-    def func(self, params):
-        return np.sum(self.wts * self.alpha * params**2)
-
-    def grad(self, params):
-        return 2 * self.wts * self.alpha * params
 
 
 
-class LogitGam(PenalizedMixin, Logit):
-    pass
 
 
 
-class LinearGam(PenalizedMixin, RLM):
-    pass
-
-n = 200
-data = pd.DataFrame()
-x = np.linspace(-10, 10, n)
-poly = x*x 
-y = 1/(1 + np.exp(-poly))
-y[y>0.5] = 1
-y[y<=0.5] = 0 
-d = {"x": x}
-dm = dmatrix("bs(x, df=5, degree=2, include_intercept=True)", d)
 
 
-   
-
-df = 10
-degree = 5
-order = degree + 1
-n_inner_knots = df - order
-lower_bound = np.min(x)
-upper_bound = np.max(x)
-knot_quantiles = np.linspace(0, 1, n_inner_knots + 2)[1:-1]
-inner_knots = _R_compat_quantile(x, knot_quantiles)
-all_knots = np.concatenate(([lower_bound, upper_bound] * order, inner_knots))
-
-basis, der_basis, der2_basis = _eval_bspline_basis(x, all_knots, degree)
-cov_der2 = np.dot(der2_basis.T, der2_basis)
-
-params0 = np.random.normal(0, 1, df)
-
-gp = GamPenalty(wts=1, alpha=0, cov_der2=cov_der2, der2=der2_basis)    
-g = LogitGam(y, basis, penal=gp)
-res_g = g.fit()
-
-plt.plot(np.dot(basis, res_g.params))
-plt.show()
+
