+++ statsmodels/sandbox/gam_gsoc2015/gam_example.R
