+++ statsmodels/sandbox/gam_gsoc2015/smooth_basis.py
